LayoutTests/inspector/editor/text-editor-selection-to-search-expected.txt

Tests synchronizing the search input field to the editor selection.


Running: testWordSelectionToSearchQuery
return
return
